Fix overflow scrolls behavior
