+++ src/views/PdfViewer.tsx
@@ -72,7 +72,7 @@ export function PdfViewer({ file, pdfViewerRef }: PdfViewerProps) {
   return (
     <div className="pdf-viewer flex h-full w-full flex-col" ref={containerRef}>
       <Document
-        className="flex-1 overflow-scroll"
+        className="flex-1 overflow-y-auto"
         externalLinkTarget="_blank"
         file={{ data: loadFileState.data }}
         onLoadSuccess={onDocumentLoadSuccess}
